fix(model): transformers>=5 vision_tower return + float32 actions
Two fixes that let VLA ablation training run end-to-end under transformers 5.8.0:
- roboqwen25vl: vision_tower now returns BaseModelOutputWithPooling; take
  .pooler_output (post-merger image tokens), with getattr fallback for older
  transformers that return a bare tensor. Two call sites.
- libero_hdf5_dataset: yield actions as float32 (was float64); float64 propagated
  into the loss and broke backward (RuntimeError: Found dtype Double but expected
  Float). Q99 math stays float64; only the yield boundary downcasts.

+++ vlm4vla/data/libero_hdf5_dataset.py
@@ -130,7 +130,10 @@ class LiberoHDF5Dataset(IterableDataset):
                         actions, self.q01, self.q99, self.norm_mask)
                     yield {
                         "language": str(lang),
-                        "actions": actions,
+                        # float32 to match the original RLDS contract; float64
+                        # would propagate into the loss and break backward
+                        # (RuntimeError: Found dtype Double but expected Float).
+                        "actions": actions.astype(np.float32),
                         "images": agent.astype(np.uint8),
                         "gripper_images": wrist.astype(np.uint8),
                     }
